Move modbus base entity to separate module
