+++ homeassistant/components/modbus/entity.py
