+++ homeassistant/components/modbus/sensor.py
@@ -27,8 +27,8 @@ from homeassistant.helpers.update_coordinator import (
 )
 
 from . import get_hub
-from .base_platform import BaseStructPlatform
 from .const import CONF_SLAVE_COUNT, CONF_VIRTUAL_COUNT
+from .entity import BaseStructPlatform
 from .modbus import ModbusHub
 
 _LOGGER = logging.getLogger(__name__)
